+++ Runtime/Core/Http/IHttpClient.cs
@@ -6,7 +6,6 @@ using AccelByte.Models;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
 
 namespace AccelByte.Core {
     public enum HttpAuth